src/Diamond/DiamondMaker.cs
public class DiamondMaker {
  public static readonly char[] alphabet = Enumerable.Range('A', 26).Select(x => (char) x).ToArray();

  public static string GenerateDiamond(char midLetter) {
    // The problem definition doesn't specify what to do with letters outside the
    // roman alphabet, so we'll assume them to be invalid.
    if (!alphabet.Contains(midLetter))
      throw new ArgumentException($"Invalid character: {midLetter}. Accepted characters are {new string(alphabet)}", nameof(midLetter));

    // The position of the character in the alphabet let us know how many lines we need to print.
    var pos = Array.IndexOf(alphabet, midLetter);

    // We can then iterate over the characters up to the position of the char we want to display.
    var lines = Enumerable.Range(0, pos + 1).Select(i => {
      var c = alphabet[i];

      // There are some patterns we can observe in the output, regarding
      // the number of spaces we need to print before and after the letter.
      var leadingSpaces = new string(' ', Math.Max(pos - i, 0));
      var trailingSpaces = new string(' ', Math.Max(i * 2 - 1, 0));

      // The first line is a special case where the letter only appears once. 
      var line = i != 0 ? $"{leadingSpaces}{c}{trailingSpaces}{c}" : $"{leadingSpaces}{c}";

      // NOTE: The attentive reader will notice that one could also easily append the reversed line,
      // albeit at the cost of generating trailing spaces that we would need to trim (or not?) later.
      // We'll leave that approach as something to be considered depending on further clarification
      // of requirements (as well as the usage of a StringBuilder).

      return line.ToString();
    });

    // Once we computed everything up to the middle line, we can just append the lines in reverse order,
    // that is, without the duplicated middle line, obviously.
    var resultLines = lines.Concat(lines.Reverse().Skip(1));
    return string.Join(Environment.NewLine, resultLines);
  }
}